fix (backend): switch to Ticker.history() and fix crypto price resolution
Replaces yf.download() with yf.Ticker.history() to avoid multi-level
column issues in newer yfinance versions. Fixes _resolve_price to not
double-append currency suffix when the ticker already contains one
(e.g. BTC-USD). Applies the same change to FX rate fetching.

+++ api/adapters/outbound/market_data/price_cache.py
@@ -1,24 +1,23 @@
-from decimal import Decimal
 from datetime import datetime, timedelta
 from typing import Optional, Dict, Tuple
 
 class PriceCache:
     def __init__(self, ttl_seconds: int = 3600):
         self.ttl = ttl_seconds
-        self._cache: Dict[str, Tuple[Decimal, datetime]] = {}
-    
-    def get(self, ticker: str) -> Optional[Decimal]:
+        self._cache: Dict[str, Tuple[int, datetime]] = {}
+
+    def get(self, ticker: str) -> Optional[int]:
         if ticker not in self._cache:
             return None
-        
+
         price, timestamp = self._cache[ticker]
         if datetime.utcnow() - timestamp > timedelta(seconds=self.ttl):
             del self._cache[ticker]
             return None
-        
+
         return price
-    
-    def set(self, ticker: str, price: Decimal) -> None:
+
+    def set(self, ticker: str, price: int) -> None:
         self._cache[ticker] = (price, datetime.utcnow())
     
     def invalidate(self, ticker: str) -> None:
